Agrego swagger y otras yerbitas

+++ src/main/java/com/grupo6/dssd/repository/ProtocolRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.grupo6.dssd.model.Protocol;
 
 @Repository
-public interface ProtocolRepository extends JpaRepository<Protocol, Integer> {
+public interface ProtocolRepository extends JpaRepository<Protocol, Long> {
 
 }
